reorganize universal system object tiers into folder
