+++ CompilerSource/compiler/components/write_object_data.cpp
@@ -666,8 +666,8 @@ static inline void write_object_declarations(lang_CPP* lcpp, const GameData &gam
   ofstream wto;
   wto.open((codegen_directory + "Preprocessor_Environment_Editable/IDE_EDIT_objectdeclarations.h").c_str(),ios_base::out);
   wto << license;
-  wto << "#include \"Universal_System/collisions_object.h\"\n";
-  wto << "#include \"Universal_System/object.h\"\n\n";
+  wto << "#include \"Universal_System/Object_Tiers/collisions_object.h\"\n";
+  wto << "#include \"Universal_System/Object_Tiers/object.h\"\n\n";
   wto << "#include <map>";
 
   declare_scripts(wto, game, state);
