+++ ENIGMAsystem/SHELL/Collision_Systems/BBox/BBOXfuncs.cpp
@@ -19,7 +19,7 @@
 // GM front-end functions - Implementations of standard GM collision functions.
 ////////////////////////////////////
 
-#include "Universal_System/collisions_object.h"
+#include "Universal_System/Object_Tiers/collisions_object.h"
 #include "Universal_System/instance_system.h" //iter
 #include "Universal_System/roomsystem.h"
 #include "Collision_Systems/collision_mandatory.h" //iter
