+++ ENIGMAsystem/SHELL/Collision_Systems/BBox/BBOXimpl.h
@@ -15,7 +15,7 @@
 *** with this code. If not, see <http://www.gnu.org/licenses/>
 **/
 
-#include "Universal_System/collisions_object.h"
+#include "Universal_System/Object_Tiers/collisions_object.h"
 
 enigma::object_collisions* const collide_inst_inst(int object, bool solid_only, bool notme, double x, double y);
 enigma::object_collisions* const collide_inst_rect(int object, bool solid_only, bool notme, int x1, int y1, int x2, int y2);
