+++ ENIGMAsystem/SHELL/Collision_Systems/Precise/PRECimpl.cpp
@@ -21,7 +21,7 @@
 // In this case, we treat instances as their bounding box (BBox).
 ////////////////////////////////////
 
-#include "Universal_System/collisions_object.h"
+#include "Universal_System/Object_Tiers/collisions_object.h"
 #include "Universal_System/instance_system.h" //iter
 #include "Universal_System/instance.h"
 #include "Universal_System/math_consts.h"
@@ -875,4 +875,3 @@ void change_inst_point(int obj, bool perf, int x1, int y1)
         }
     }
 }
-
