+++ ENIGMAsystem/SHELL/ENIGMAengine.cbp
@@ -953,8 +953,8 @@
 		<Unit filename="Universal_System/bufferstruct.h" />
 		<Unit filename="Universal_System/callbacks_events.cpp" />
 		<Unit filename="Universal_System/callbacks_events.h" />
-		<Unit filename="Universal_System/collisions_object.cpp" />
-		<Unit filename="Universal_System/collisions_object.h" />
+		<Unit filename="Universal_System/Object_Tiers/collisions_object.cpp" />
+		<Unit filename="Universal_System/Object_Tiers/collisions_object.h" />
 		<Unit filename="Universal_System/darray.cpp" />
 		<Unit filename="Universal_System/darray.h" />
 		<Unit filename="Universal_System/debugscope.cpp" />
@@ -972,8 +972,8 @@
 		<Unit filename="Universal_System/fontstruct.h" />
 		<Unit filename="Universal_System/globalupdate.cpp" />
 		<Unit filename="Universal_System/globalupdate.h" />
-		<Unit filename="Universal_System/graphics_object.cpp" />
-		<Unit filename="Universal_System/graphics_object.h" />
+		<Unit filename="Universal_System/Object_Tiers/graphics_object.cpp" />
+		<Unit filename="Universal_System/Object_Tiers/graphics_object.h" />
 		<Unit filename="Universal_System/highscore_functions.cpp" />
 		<Unit filename="Universal_System/highscore_functions.h" />
 		<Unit filename="Universal_System/image_formats.cpp" />
@@ -1009,12 +1009,12 @@
 		<Unit filename="Universal_System/multifunction_variant.cpp" />
 		<Unit filename="Universal_System/multifunction_variant.h" />
 		<Unit filename="Universal_System/nlpo2.h" />
-		<Unit filename="Universal_System/object.cpp" />
-		<Unit filename="Universal_System/object.h" />
+		<Unit filename="Universal_System/Object_Tiers/object.cpp" />
+		<Unit filename="Universal_System/Object_Tiers/object.h" />
 		<Unit filename="Universal_System/path_functions.h" />
 		<Unit filename="Universal_System/pathstruct.h" />
-		<Unit filename="Universal_System/planar_object.cpp" />
-		<Unit filename="Universal_System/planar_object.h" />
+		<Unit filename="Universal_System/Object_Tiers/planar_object.cpp" />
+		<Unit filename="Universal_System/Object_Tiers/planar_object.h" />
 		<Unit filename="Universal_System/rectpack.cpp" />
 		<Unit filename="Universal_System/rectpack.h" />
 		<Unit filename="Universal_System/reflexive_types.cpp" />
@@ -1036,10 +1036,10 @@
 		<Unit filename="Universal_System/syntax_quirks.h" />
 		<Unit filename="Universal_System/terminal_io.cpp" />
 		<Unit filename="Universal_System/terminal_io.h" />
-		<Unit filename="Universal_System/timelines_object.cpp" />
-		<Unit filename="Universal_System/timelines_object.h" />
-		<Unit filename="Universal_System/transform_object.cpp" />
-		<Unit filename="Universal_System/transform_object.h" />
+		<Unit filename="Universal_System/Object_Tiers/timelines_object.cpp" />
+		<Unit filename="Universal_System/Object_Tiers/timelines_object.h" />
+		<Unit filename="Universal_System/Object_Tiers/transform_object.cpp" />
+		<Unit filename="Universal_System/Object_Tiers/transform_object.h" />
 		<Unit filename="Universal_System/var4.cpp" />
 		<Unit filename="Universal_System/var4.h" />
 		<Unit filename="Universal_System/var4_lua.cpp" />
