+++ ENIGMAsystem/SHELL/Graphics_Systems/General/GSscreen.cpp
@@ -28,7 +28,7 @@
 #include "GScolors.h"
 
 #include "Universal_System/background.h"
-#include "Universal_System/graphics_object.h"
+#include "Universal_System/Object_Tiers/graphics_object.h"
 #include "Universal_System/depth_draw.h"
 #include "Universal_System/instance_system.h"
 #include "Universal_System/roomsystem.h"
