+++ ENIGMAsystem/SHELL/Universal_System/Extensions/Paths/path_functions.cpp
@@ -47,7 +47,7 @@ namespace enigma {
 	extern unsigned bound_texture;
 }
 
-#include "Universal_System/collisions_object.h"
+#include "Universal_System/Object_Tiers/collisions_object.h"
 #include "Universal_System/instance_system.h"
 #include "implement.h"
