+++ ENIGMAsystem/SHELL/Universal_System/Makefile
@@ -1,3 +1,4 @@
-SOURCES += $(wildcard Universal_System/*.cpp)
+SOURCES += $(wildcard Universal_System/*.cpp) $(wildcard Universal_System/Object_Tiers/*.cpp)
 override LDLIBS += -lz
+override CXXFLAGS += -IUniversal_System/ -IUniversal_System/Object_Tiers
 CLEANDEPS += $(OBJDIR)/Universal_System/loading.o
