+++ ENIGMAsystem/SHELL/Universal_System/Object_Tiers/collisions_object.cpp
