+++ ENIGMAsystem/SHELL/Universal_System/Object_Tiers/graphics_object.h
