+++ ENIGMAsystem/SHELL/Universal_System/Object_Tiers/object.h
