+++ ENIGMAsystem/SHELL/Universal_System/Object_Tiers/planar_object.h
