+++ ENIGMAsystem/SHELL/Universal_System/Object_Tiers/transform_object.cpp
