+++ ENIGMAsystem/SHELL/Widget_Systems/DlgMod/dialogs.cpp
@@ -22,7 +22,7 @@
 #ifdef DEBUG_MODE
 #include "Universal_System/var4.h"
 #include "Universal_System/resource_data.h"
-#include "Universal_System/object.h"
+#include "Universal_System/Object_Tiers/object.h"
 #include "Universal_System/debugscope.h"
 #endif
