+++ ENIGMAsystem/SHELL/Widget_Systems/Win32/dialogs.cpp
@@ -48,7 +48,7 @@ static string gs_but1, gs_but2, gs_but3;
 #ifdef DEBUG_MODE
 #include "Universal_System/var4.h"
 #include "Universal_System/resource_data.h"
-#include "Universal_System/object.h"
+#include "Universal_System/Object_Tiers/object.h"
 #include "Universal_System/debugscope.h"
 #endif
